tasks/5_tax_year_assessments_config/requirements.txt
python-dotenv==1.0.1
functions_framework==3.8.2
protobuf==6.30.2
pyproj==3.7.1
Shapely==2.0.7
google-cloud-storage==2.16.0